drivers: video: msm_mdp: update userspace change max Hz
While before there was only the on / off toggle, now there is the possibility to choose more frequencies.
So we go from a bool to an int. For now the frequencies are restricted to 3 options:

- Default (if developer do OC screen)
- 60 Hz
- 55 Hz
- 50 Hz

here: /sys/module/mdss_mdp/parameters/custom_hz

NB: Even if you change, you will find the default Hz in the HW information

Signed-off-by: Peppe289 <gsperanza204@gmail.com>